fix ie8 onclick bug...
In IE8 the jQuery event object for onclick has the "which" property always set
to 0. Previously we were relying on it being 1 to indiciate a left mouse click.
Also, it appears that right mouse clicks never fire the click event so not sure
why we needed that code in the first place. Now it works in ie8 again without
causing a page refresh.